Moved sample analyser to database. Adding sample item to database when adding files. Adding sample tags to sample item when adding files.

+++ Saempl/Source/SampleAnalyser.cpp
@@ -20,7 +20,7 @@ SampleAnalyser::~SampleAnalyser()
     
 }
 
-float SampleAnalyser::analyseSampleLength(File inFile)
+double SampleAnalyser::analyseSampleLength(File inFile)
 {
     loadAudioFileSource(inFile);
